feat(chunk_delta_h): output ht as fp32 via direct reg→GMEM store
- Remove TMA path for ht (bf16 SMEM→GMEM)
- Compute warp writes fp32 directly from registers to GMEM
  using per-element coordinate loop (like KDA kernel)
- Preserves fp32 accumulator precision, aligns with FLA output dtype
- All callers allocate fp32 ht tensors
- 46/46 tests pass, ht error reduced (e.g. 0.0040→0.0033)
- Performance unchanged: bench_h 1.70x, bench_varlen 1.34x

+++ flashla/bench_h_kernel.py
@@ -81,7 +81,7 @@ def run_benchmark(B, T, H, K, V, BT, use_g, use_gk, use_h0, store_ht, save_vnew)
 
     h_out = torch.zeros(B, NT, H, K, V, device=device, dtype=dtype)
     v_new_out = torch.zeros(B, T, H, V, device=device, dtype=dtype)
-    ht_out = torch.zeros(B, H, K, V, device=device, dtype=dtype)
+    ht_out = torch.zeros(B, H, K, V, device=device, dtype=torch.float32)
 
     kernel = ChunkDeltaRuleFwdH(chunk_size=BT, head_dim_k=K, head_dim_v=V)
     stream = cutlass_torch.default_stream()
