+++ flashla/bench_varlen.py
@@ -141,7 +141,7 @@ def run_varlen_benchmark(num_seqs, total_T, H, K, V, BT, ratio,
     
     h_out = torch.zeros(1, total_NT, H, K, V, device=device, dtype=dtype)
     v_new_out = torch.zeros(1, total_T, H, V, device=device, dtype=dtype)
-    ht_out = torch.zeros(num_seqs, H, K, V, device=device, dtype=dtype)
+    ht_out = torch.zeros(num_seqs, H, K, V, device=device, dtype=torch.float32)
     workspace = torch.zeros(num_seqs * 128, dtype=torch.uint8, device=device)
     
     kernel = ChunkDeltaRuleFwdH(chunk_size=BT, head_dim_k=K, head_dim_v=V, is_varlen=True, persistent=persistent)
